conversational_service/requirements.txt
# Conversational Recommendation Service Requirements

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9

# LLM Integration (Gemini only - FREE tier!)
google-genai>=0.2.0

# Environment & Config
python-dotenv==1.0.1
pyyaml==6.0.2

# Data Processing
pandas==2.2.3
numpy==2.1.3

# Vector Search (FAISS compatibility)
faiss-cpu>=1.9.0

# CLIP for text encoding
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0

# HTTP Client
httpx==0.27.2
requests==2.32.3

# Utilities
python-json-logger==2.0.7
